+++ CHANGELOG.md
@@ -1,19 +1,23 @@
 CHANGELOG
 =========
 
-Version 0.4.1 - Beta
---------------------
+Unreleased Changes!
+-------------------
 -   Add: The new Arabic Translation. Many Thanks to [Mohammed Alsiddeeq Ahmed](https://github.com/mosid)
          [#1](https://github.com/pytesNET/tail.DateTime/issues/1)!
 -   Add: The new option `rtl`, which allows to display the DateTime picker in an RTL style.
+-   Add: The new event `view`.
 -   Update: The new RTL supported Stylesheets.
 -   Update: The default tooltip tick color depends now on the used theme.
 -   Update: The "Enter/Escape" Key listener now also works when the cursor is within an input field.
+-   Update: The events now gets fired AFTER the DateTime picker has done his actions!
 -   Bugfix: Disabled dates could still be selected / clicked.
 -   Bugfix: The `keyup` document event listener has fired per instance.
 -   Bugfix: The "Enter" Key listener doesn't notice / used already selected dates!
 -   Bugfix: You cannot go back farther than to February of the displayed year.
             [#3](https://github.com/pytesNET/tail.DateTime/issues/3)!
+-   Bugfix: Issue with Date Selection with Update to 0.4.0.
+            [#2](https://github.com/pytesNET/tail.DateTime/issues/2)!
 
 Version 0.4.0 - Beta
 --------------------